+++ threedigrid_builder/tests/test_connection_nodes.py
@@ -6,7 +6,6 @@ from threedigrid_builder.constants import CalculationType
 from threedigrid_builder.constants import ContentType
 from threedigrid_builder.constants import LineType
 from threedigrid_builder.constants import NodeType
-from threedigrid_builder.exceptions import SchematisationError
 from threedigrid_builder.grid import Channels
 from threedigrid_builder.grid import ConnectionNodes
 from threedigrid_builder.grid import CrossSectionLocations
@@ -183,12 +182,13 @@ def test_set_bottom_levels_multiple_nodes(_type):
 
 
 @pytest.mark.parametrize("structure_type", [Weirs, Orifices, Pipes, Culverts, Channels])
-def test_bottom_levels_above_invert_level(structure_type):
+def test_bottom_levels_above_invert_level(structure_type, caplog):
     nodes = Nodes(
         id=[1, 2],
         content_type=ContentType.TYPE_V2_CONNECTION_NODES,
         dmax=[10.0, 20.0],
         content_pk=[52, 22],
+        manhole_id=[3, 6],
     )
     lines = Lines(
         id=[1],
@@ -196,13 +196,15 @@ def test_bottom_levels_above_invert_level(structure_type):
         content_pk=2,
         line=[[1, 2]],
         invert_level_start_point=3.0,
-        invert_level_end_point=3.0,
+        invert_level_end_point=20.0,
     )
 
+    # setting the bottom levels emits a warning
+    set_bottom_levels(nodes, lines)
+    assert caplog.messages[0].startswith("Manholes [3] have a bottom_level")
+
     # assert the resulting value of dmax
-    with pytest.raises(SchematisationError) as e:
-        set_bottom_levels(nodes, lines)
-        assert str(e).startswith("Connection nodes [22, 52] have a manhole")
+    assert_almost_equal(nodes.dmax, [3.0, 20.0])
 
 
 def test_1d2d_properties(connection_nodes):