+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog of threedigrid-builder
 
 - Dropped support for Python 3.6
 
+- Fixed __version__ attribute and  "threedigrid_builder_version" HDF5 attribute.
+
 
 0.8.0 (2021-11-30)
 ------------------